% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta2way.R
\name{meta2way.list}
\alias{meta2way.list}
\title{Two-way meta-analysis of bootstrapped effect and cost data}
\usage{
\method{meta2way}{list}(x.bt, y.bt, type = "both", study.names = NULL)
}
\description{
For list objects
}
