% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta2way.R
\name{meta2way}
\alias{meta2way}
\title{Two-way meta-analysis of bootstrapped effect and cost data}
\usage{
meta2way(...)
}
\arguments{
\item{x.bt}{list of each studies x axis bootstrap values.}

\item{y.bt}{list of each studies y axis bootstrap values.}

\item{type}{character which type of meta-analysis to perform. Acceptes either of 'fixed', 'random' or 'both'. Defaults to 'both'.}

\item{study.names}{character with readable names of the studies. Defaults to NULL, in which case the names of x.bt will be used (with a warning if names of y.bt are different or not available). If x.bt has no names, 'Study A','Study B', etc will be used.}
}
\value{
meta2way returns an object of class "meta2way" which represents a two-way meta-analysis of studies, e.g. cost-effectiveness studies.
The function plot can be used to plot the two dimensions of the meta-analysis, such as a cost-effectiveness plane. The function print will display the summary results of each study and the meta-analysis for the two dimentions. The function willing2pay can be used to plot a willingness to pay curve.
An object of class "meta2way" is a named list containing the following components:
\describe{
  \item{x}{named list with values for x-variable:
    \describe{
      \item{bt}{copy of x.bt; if study.names was provided, these will be used as names of bt}
      \item{meta.fixed}{fixed effects meta analysis result for x}
      \item{meta.random}{random effects meta-analysis result for x}
      \item{Q}{value of Q for x}
      \item{tau2}{value of tau2 for x}
      \item{I2}{value of I2 for x}
    }}
  \item{y}{named list with values for y-variable, same as x}
  \item{study.names}{character vector of study names}
  \item{do.fixed}{logical whether or not fixed effects meta-analysis is performed}
  \item{do.random}{logical whether or not random effects meta-analysis is performed}
  \item{n.studies}{integer length 1 containing number of studies}
  \item{n.bt}{integer length 1 containing number of bootstrap samples per study}
}
}
\description{
This function allows you to perform a two-way meta-analysis of bootstrapped effect and cost data
}
\examples{
# perform meta-analysis using list of x-values and list of y-values
m <- meta2way(list(Abrahams=treatments$Abrahams2018.x, Baruch=treatments$Baruch2018.x),
              list(Abrahams=treatments$Abrahams2018.y, Baruch=treatments$Baruch2018.y))
m
# this is a trick to do meta-analysis of the inverse comparison
m <- meta2way(list(Abrahams=-treatments$Abrahams2018.x, Baruch=-treatments$Baruch2018.x),
              list(Abrahams=-treatments$Abrahams2018.y, Baruch=-treatments$Baruch2018.y))
m

# meta-analysis can also be performed using study2way objects
s1 <- study2way(treatments$Abrahams2018.x, treatments$Abrahams2018.y, 'Abrahams 2018')
s2 <- study2way(treatments$Baruch2018.x, treatments$Baruch2018.y, 'Baruch 2018')
m <- meta2way(s1, s2)
m
}
\keyword{Cost-effectiveness}
\keyword{Two-way}
\keyword{meta-analysis;}
