% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta2way.R
\name{meta2way.study2way}
\alias{meta2way.study2way}
\title{Two-way meta-analysis of bootstrapped effect and cost data}
\usage{
\method{meta2way}{study2way}(..., type = "both")
}
\description{
For the study2way objects
}
