% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.meta2way.R
\name{plot.meta2way}
\alias{plot.meta2way}
\title{Plot two-way meta-analysis cost-effectiveness plane based on bootstrap data}
\usage{
\method{plot}{meta2way}(x, y, xlab = "x", ylab = "y", col = NULL,
  pch = NULL, lwd = 1, lty = 1, points = TRUE, meta.col = 1,
  meta.pch = c(17, 15), meta.lwd = 2, meta.lty = c(2, 3),
  meta.points = FALSE, max.points = 2000,
  quadrant.proportions = FALSE, legend = TRUE, legend.ncol = 2,
  legend.cex = 1)
}
\arguments{
\item{x}{Object of class meta2way.}

\item{y}{Should be omitted.}

\item{xlab}{x axis label. Defaults to 'x'.}

\item{ylab}{y axis label. Defaults to 'y'.}

\item{col}{color for individual studies. Defaults to 2:(x$n.studies+1).}

\item{pch}{integer point character for individual studies bootstrap points and confidence area. Length should equal number of included studies. Defaults to 2:(x$n.studies+1).}

\item{lwd}{integer line width for individual studies confidence area. Defaults to 1.}

\item{points}{logical whether bootstrap points of individual studies should be plotted. Defaults to TRUE.}

\item{meta.col}{color for meta-analysis bootstrap points and confidence area. Vector length one or two; if two, the first will be used for the fixed effects and the second for the random effects meta-analysis if performed. Defaults to 1.}

\item{meta.pch}{integer point character for meta-analysis points: vector length one or two (like meta.col). Defaults to c(15,17) because I like these characters most.}

\item{meta.lwd}{integer line width for meta-analysis confidence area. Defaults to 2.}

\item{meta.lty}{integer line type for meta-analysis confidence area. Defaults to c(2,3).}

\item{meta.points}{logical whether bootstrap points of meta-analyses should be plotted. Defaults to FALSE.}

\item{max.points}{integer the maximum number of bootstrap points to be plotted per study and meta-analysis. Bootstrap samples 1 to max.points will be plotted if available. Use NA or NULL to plot all points. Defaults to 2000.}

\item{quadrant.proportions}{logical whether the proportions of meta-analysis bootstrap points in each quadrant should be printed in the plot corners. Defaults to FALSE.}

\item{legend}{logical whether a legend of the studies should be printed. The legend will be printed below the plot. Defaults to TRUE.}

\item{legend.ncol}{integer number of columns of the legend. Defaults to 2.}

\item{legend.cex}{relative font size of the legend. Defaults to 1.}
}
\value{
Returns NULL invisibly.
}
\description{
This function plots the cost-efectiveness plane of a two-way meta-analysis.
}
\examples{
m <- meta2way(study2way(treatments$Abrahams2018.x, treatments$Abrahams2018.y, 'Abrahams 2018'),
              study2way(treatments$Baruch2018.x, treatments$Baruch2018.y, 'Baruch 2018'))
plot(m, xlab='delta effects', ylab='delta costs')
}
\keyword{bootstrapping}
\keyword{cost-effectiveness;}
\keyword{meta-analysis;}
