% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.study2way.R
\name{plot.study2way}
\alias{plot.study2way}
\title{Plot two-way analysis cost-effectiveness plane based on bootstrap data}
\usage{
\method{plot}{study2way}(x, y, xlab = "x", ylab = "y",
  xlim = range(x$x.bt), ylim = range(x$y.bt), col = 1, pch = 3,
  center.pch = 15, lwd = 1, lty = 1, points = TRUE,
  max.points = 2000, quadrant.proportions = FALSE, add = FALSE)
}
\arguments{
\item{x}{Object of class study2way.}

\item{y}{Should be omitted.}

\item{xlab}{x axis label. Defaults to 'x'.}

\item{ylab}{y axis label. Defaults to 'y'.}

\item{xlim}{numeric vector length 2 with limits for the x axis. Defaults to range of x$x.bt.}

\item{ylim}{numeric vector length 2 with limits for the y axis. Defaults to range of x$y.bt.}

\item{col}{color for individual studies. Defaults to 1.}

\item{pch}{integer point character for bootstrap points and confidence area. Defaults to 2.}

\item{lwd}{integer line width for individual studies confidence area. Defaults to 1.}

\item{points}{logical whether bootstrap points should be plotted. Defaults to TRUE.}

\item{max.points}{integer the maximum number of bootstrap points to be plotted per study and meta-analysis. Bootstrap samples 1 to max.points will be plotted if available. Use NA or NULL to plot all points. Defaults to 2000.}

\item{quadrant.proportions}{logical whether the proportions of bootstrap points in each quadrant should be printed in the plot corners. Defaults to FALSE.}

\item{add}{logical whether to add the plot to existing plot. Defaults to FALSE.}
}
\value{
Returns NULL invisibly.
}
\description{
This function plots the cost-efectiveness plane of a two-way analysis of a single study.
}
\examples{
s1 <- study2way(treatments$Abrahams2018.x, treatments$Abrahams2018.y, 'Abrahams 2018')
plot(s1, xlab='Delta effects', ylab='Delta costs')
}
\keyword{bootstrapping}
\keyword{cost-effectiveness;}
