% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.meta2way.R
\name{print.meta2way}
\alias{print.meta2way}
\title{Print the two-way meta-analysis results based on bootstrap data}
\usage{
\method{print}{meta2way}(object)
}
\arguments{
\item{object}{Object of class meta2way.}
}
\value{
Returns NULL invisibly.
}
\description{
This function allows you to print the two-way meta-analysis results based on bootstrap data.
}
\examples{
m <- meta2way(study2way(treatments$Abrahams2018.x, treatments$Abrahams2018.y, 'Abrahams 2018'),
              study2way(treatments$Baruch2018.x, treatments$Baruch2018.y, 'Baruch 2018'))
m
print(m)
}
\keyword{bootstrapping}
\keyword{cost-effectiveness;}
\keyword{meta-analysis;}
