% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study2way.R
\name{study2way}
\alias{study2way}
\title{Two-way analysis of bootstrapped effect and cost data}
\usage{
study2way(x.bt, y.bt, study.name)
}
\arguments{
\item{x.bt}{vector of the x axis bootstrap values.}

\item{y.bt}{vector of the y axis bootstrap values.}

\item{study.name}{character with readable name of the studies.}
}
\value{
study2way returns an object of class "study2way" which represents a two-way analysis of studies, e.g. cost-effectiveness studies.
The function plot can be used to plot the two dimensions of the analysis, such as a cost-effectiveness plane. The function print will display the results of the study for the two dimentions. The function willing2pay can be used to plot a willingness to pay curve.
An object of class "study2way" is a named list containing the following components:
\describe{
  \item{x.bt}{copy of x.bt}
  \item{y.bt}{copy of y.bt}
  \item{study.name}{character vector of study name}
  \item{n.bt}{integer length 1 containing number of bootstrap samples per study}
}
}
\description{
This function allows you to perform a two-way analysis of bootstrapped effect and cost data of a single study
}
\examples{
# plot two-way analysis of study 1
s1 <- study2way(treatments$Abrahams2018.x, treatments$Abrahams2018.y, 'Abrahams 2018')
plot(s1)

# Add study 2 and perform meta-analysis
s2 <- study2way(treatments$Baruch2018.x, treatments$Baruch2018.y, 'Baruch 2018')
m <- meta2way(s1, s2)
plot(m)
}
\keyword{Cost-effectiveness}
\keyword{Two-way}
\keyword{meta-analysis;}
