% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatments}
\alias{treatments}
\title{difference in effects and costs of two treatments.}
\format{A data frame with 2,000 rows and 6 columns:
\describe{
  \item{Abrahams2018.x}{x values (effects) of the bootstrap results of the first study (Abrahams 2018)}
  \item{Abrahams2018.y}{y values (costs) of the bootstrap results of the first study (Abrahams 2018)}
  \item{Baruch2018.x}{x values of second study}
  \item{Baruch2018.y}{y values of second study}
  \item{Chananja2018.x}{x values of third study}
  \item{Chananja2018.y}{y values of third study}
}}
\usage{
treatments
}
\description{
A data frame containing 2,000 bootstrap results of cost-effectiveness analysis of three fictitious studies.
}
\keyword{datasets}
