% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/willing2pay.R
\name{willing2pay}
\alias{willing2pay}
\title{Willingness to pay plot}
\usage{
willing2pay(object, type, xlim = NA, xlim.quantile = 0.999,
  ylim = 0:1, xlab = "Costs per effect",
  ylab = "Probability of acceptance", main = "Willingness to Pay",
  col = 1, lty = 1, add = FALSE, res = 1000)
}
\arguments{
\item{object}{object of class meta2way.}

\item{type}{character length 1, either 'fixed' or 'random'. If object contains only one type of meta-analysis, this argument can be omitted.}

\item{xlim}{numeric length 2 with the x-axis limits. Defaults to NA which means the argument xlim.quantile is used to determine xlim.}

\item{xlim.quantile}{numeric length 1: which upper extreme of costs per effect will be plotted. Defaults to .999. Setting the value to 1 means the most extreme value will be plotted.}

\item{ylim}{numeric length 2 with the y-axis limits. Defaults to c(0,1)}

\item{xlab}{character x-axis label. Defaults to 'Costs per effect'}

\item{ylab}{character y-axis label. Defaults to 'Probability of treshold'}

\item{main}{character plot title. Defaults to 'Willingness to Pay'}

\item{col}{line color. Defaults to 1.}

\item{lty}{line type. Defaults to 1.}

\item{add}{logical whether to add the line to the previous plot. Defaults to FALSE.}

\item{res}{integer how many data points should be calculated between xlim[1] and xlim[2]. Defaults to 1000.}
}
\value{
Returns NULL invisibly.
}
\description{
This function allows you to plot a willingness to pay plot based on a two-way meta analysis object.
}
\examples{
m <- meta2way(study2way(treatments$Abrahams2018.x, treatments$Abrahams2018.y, 'Abrahams 2018'),
              study2way(treatments$Baruch2018.x, treatments$Baruch2018.y, 'Baruch 2018'))
willing2pay(m, 'fixed')
}
\keyword{Willingness}
\keyword{cost-effectiveness;}
\keyword{meta-analysis}
\keyword{pay;}
\keyword{to}
